requirements.txt
streamlit
langchain-xai
langchain-community
pypdf
python-dotenv
